+++ src/vtex/getShelfByCollection/getShelfByCollection.types.ts
@@ -1,0 +1,9 @@
+interface ISearchParams {
+  collectionId: string | number;
+  pageSize: number | string;
+  pageNumber: number | string;
+  shelfTemplateId: string;
+  order?: string;
+}
+
+export type { ISearchParams };