+++ src/vtex/getShelfByCollection/index.ts
@@ -1,0 +1,3 @@
+import { getShelfByCollection } from './getShelfByCollection';
+
+export default getShelfByCollection;