+++ CHANGELOG.md
@@ -10,8 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added automation tools to handle documentation.
+- [vtex] Added `getShelfByCollection` method.
 - [vtex] Added `getShelfById` method.
 - [vtex] Added `getShelfBySku` method.
+
+### Changed
+
+- [vtex] Deprecated `getCollectionShelf` method. Use `getShelfByCollection` instead.
 
 ## [1.0.0] - 06-04-2022